Fix cpy
Fix the list generated when copying text. Previously it didn't link correctly. Remaining issue regarding copy is to chain this new list to the main list when doing a paste, thus pasting the actual text. Other change is to add template functions for moving parts not related to modification of the linked lists to editorMode.h/c.